sources/Project/requirements.txt
wheel~=0.34.2
psycopg2-binary~=2.8.5
python-dateutil~=2.8.0
SQLAlchemy~=1.3.10
Flask~=1.0.2
flasgger~=0.9.2
flask_caching~=1.8.0
flask_cors~=3.0.8
python-dateutil~=2.8.0
pandas~=1.0.3
backoff~=1.8.0
requests~=2.23.0
beautifulsoup4~=4.7.1
alembic~=1.4.2
APScheduler~=3.6.1
python-dotenv~=0.13.0
pathlib~=1.0.1
Beaker~=1.11.0